Refactor exception handling structure
